Added a new reference
Hi, I have added a new reference Java, in this line - We introduce you to one of the most popular programming languages used in the world, Java. I was going through the article and found that more about Java will help the learners, so thought of adding it here. Hope this small contribution will help other learners.